reformatting the repo
