+++ src/task_generator.rs
@@ -9,10 +9,10 @@ use crate::task_placer::TaskPlacer;
 use crate::util::MyDurationRound;
 
 /// A range of datetimes with an interval.
-struct DateRange {
-	start: NaiveDateTime,
-	end: NaiveDateTime,
-	interval: Duration,
+pub(crate) struct DateRange {
+	pub(crate) start: NaiveDateTime,
+	pub(crate) end: NaiveDateTime,
+	pub(crate) interval: Duration,
 }
 
 impl IntoIterator for DateRange {
@@ -33,7 +33,7 @@ impl IntoIterator for DateRange {
 	}
 }
 
-struct DateRangeIter {
+pub(crate) struct DateRangeIter {
 	end: NaiveDateTime,
 	current_start: NaiveDateTime,
 	current_end: NaiveDateTime,
@@ -106,241 +106,4 @@ pub fn task_generator(Input { start, end, goals }: Input) -> TaskPlacer {
 	}
 
 	TaskPlacer::new(tasks, slots)
-}
-
-#[cfg(test)]
-mod tests {
-	use crate::goal::{Goal, Repetition};
-	use crate::input::Input;
-	use crate::task_generator::task_generator;
-
-	use super::*;
-
-	#[test]
-	fn date_range_iter_simple() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-			end: NaiveDate::from_ymd(2022, 1, 2).and_hms(0, 0, 0),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0)
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 2).and_hms(0, 0, 0),
-				),
-			]
-		)
-	}
-
-	#[test]
-	fn date_range_single() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-			end: NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![(
-				NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-				NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0)
-			),]
-		)
-	}
-
-	#[test]
-	fn date_range_single_not_round() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 0),
-			end: NaiveDate::from_ymd(2022, 1, 1).and_hms(7, 0, 0),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![(
-				NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 0),
-				NaiveDate::from_ymd(2022, 1, 1).and_hms(7, 0, 0)
-			),]
-		)
-	}
-
-	#[test]
-	fn date_range_iter_not_round_end() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-			end: NaiveDate::from_ymd(2022, 1, 1).and_hms(23, 0, 1),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(0, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0)
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(23, 0, 1),
-				),
-			]
-		)
-	}
-
-	#[test]
-	fn date_range_iter_not_round_start() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 1),
-			end: NaiveDate::from_ymd(2022, 1, 2).and_hms(0, 0, 0),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 1),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 2).and_hms(0, 0, 0),
-				),
-			]
-		)
-	}
-
-	#[test]
-	fn date_range_iter_not_round_start_end() {
-		let r = DateRange {
-			start: NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 1),
-			end: NaiveDate::from_ymd(2022, 1, 1).and_hms(23, 0, 1),
-			interval: Duration::hours(8),
-		};
-
-		assert_eq!(
-			r.into_iter().collect::<Vec<_>>(),
-			vec![
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(1, 0, 1),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(8, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-				),
-				(
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(16, 0, 0),
-					NaiveDate::from_ymd(2022, 1, 1).and_hms(23, 0, 1),
-				),
-			]
-		)
-	}
-
-	#[test]
-	fn repeat() {
-		let input = Input::new(
-			NaiveDate::from_ymd(2022, 1, 1),
-			NaiveDate::from_ymd(2022, 1, 4),
-			vec![Goal::new(1).duration(1).repetition(Repetition::DAILY)],
-		);
-
-		let scheduler = task_generator(input);
-		assert_eq!(
-			scheduler.tasks,
-			vec![Task::new(0, 1, 1), Task::new(1, 1, 1), Task::new(2, 1, 1)]
-		);
-		assert_eq!(
-			scheduler.slots,
-			vec![Slot::new(0, 0, 24), Slot::new(1, 24, 48), Slot::new(2, 48, 72)]
-		)
-	}
-
-	#[test]
-	fn repeat_with_goal_start_not_midnight() {
-		let input = Input::new(
-			NaiveDate::from_ymd(2022, 1, 1),
-			NaiveDate::from_ymd(2022, 1, 4),
-			vec![Goal::new(1)
-				.duration(1)
-				.repetition(Repetition::DAILY)
-				.start(NaiveDate::from_ymd(2022, 1, 1).and_hms(10, 0, 0))],
-		);
-
-		let scheduler = task_generator(input);
-		assert_eq!(
-			scheduler.tasks,
-			vec![Task::new(0, 1, 1), Task::new(1, 1, 1), Task::new(2, 1, 1)]
-		);
-		assert_eq!(
-			scheduler.slots,
-			vec![Slot::new(0, 10, 24), Slot::new(1, 24, 48), Slot::new(2, 48, 72)]
-		)
-	}
-
-	#[test]
-	fn repeat_with_goal_end_not_midnight() {
-		let input = Input::new(
-			NaiveDate::from_ymd(2022, 1, 1),
-			NaiveDate::from_ymd(2022, 1, 4),
-			vec![Goal::new(1)
-				.duration(1)
-				.repetition(Repetition::DAILY)
-				.deadline(NaiveDate::from_ymd(2022, 1, 3).and_hms(14, 0, 0))],
-		);
-
-		let scheduler = task_generator(input);
-		assert_eq!(
-			scheduler.tasks,
-			vec![Task::new(0, 1, 1), Task::new(1, 1, 1), Task::new(2, 1, 1)]
-		);
-		assert_eq!(
-			scheduler.slots,
-			vec![Slot::new(0, 0, 24), Slot::new(1, 24, 48), Slot::new(2, 48, 62)]
-		)
-	}
-
-	#[test]
-	fn repeat_with_goal_start_and_end_not_midnight() {
-		let input = Input::new(
-			NaiveDate::from_ymd(2022, 1, 1),
-			NaiveDate::from_ymd(2022, 1, 4),
-			vec![Goal::new(1)
-				.duration(1)
-				.repetition(Repetition::DAILY)
-				.start(NaiveDate::from_ymd(2022, 1, 1).and_hms(10, 0, 0))
-				.deadline(NaiveDate::from_ymd(2022, 1, 3).and_hms(14, 0, 0))],
-		);
-
-		let scheduler = task_generator(input);
-		assert_eq!(
-			scheduler.tasks,
-			vec![Task::new(0, 1, 1), Task::new(1, 1, 1), Task::new(2, 1, 1)]
-		);
-		assert_eq!(
-			scheduler.slots,
-			vec![Slot::new(0, 10, 24), Slot::new(1, 24, 48), Slot::new(2, 48, 62)]
-		)
-	}
 }
