+++ src/task_placer.rs
@@ -186,96 +186,3 @@ impl TaskPlacer {
 		})
 	}
 }
-
-#[cfg(test)]
-mod tests {
-	use crate::input::Input;
-	use crate::task_generator::task_generator;
-
-	#[test]
-	fn basic_test() {
-		let input: Input = serde_json::from_str(
-			r#"
-{
-    "startDate": "2022-01-01",
-    "endDate": "2022-01-02",
-    "goals": [
-        {
-          "id": 1,
-          "title" : "shopping",
-          "duration": 1,
-          "start": "2022-01-01T10:00:00",
-          "deadline": "2022-01-01T13:00:00"
-        },
-        {
-          "id": 2,
-          "title": "dentist",
-          "duration": 1,
-          "start": "2022-01-01T10:00:00",
-          "deadline": "2022-01-01T11:00:00"
-        },
-        {
-          "id": 3,
-          "title" : "exercise",
-          "duration": 1,
-          "start": "2022-01-01T10:00:00",
-          "deadline": "2022-01-01T18:00:00"
-        }
-    ]
-}
-        "#,
-		)
-		.unwrap();
-
-		let scheduler = task_generator(input);
-		let result = scheduler.task_placer().unwrap();
-		let result_json = serde_json::to_string_pretty(&result).unwrap();
-
-		//println!("{}", result_json);
-		assert_eq!(
-			result_json,
-			r#"{
-  "tasks": [
-    {
-      "id": 0,
-      "goal_id": 1,
-      "duration_to_schedule": 0,
-      "duration_scheduled": 1,
-      "status": "SCHEDULED"
-    },
-    {
-      "id": 1,
-      "goal_id": 2,
-      "duration_to_schedule": 0,
-      "duration_scheduled": 1,
-      "status": "SCHEDULED"
-    },
-    {
-      "id": 2,
-      "goal_id": 3,
-      "duration_to_schedule": 0,
-      "duration_scheduled": 1,
-      "status": "SCHEDULED"
-    }
-  ],
-  "slots": [
-    {
-      "task_id": 0,
-      "start": 11,
-      "end": 12
-    },
-    {
-      "task_id": 1,
-      "start": 10,
-      "end": 11
-    },
-    {
-      "task_id": 2,
-      "start": 13,
-      "end": 14
-    }
-  ]
-}"#
-		);
-	}
-}
